Added Treatment page and table

+++ therapy-manager/src/main/java/com/bharatonjava/therapymanager/dao/AddressDao.java
@@ -4,7 +4,7 @@ import com.bharatonjava.therapymanager.domain.Address;
 
 public interface AddressDao {
 
-	public int createAddress(Address address);
+	public Long createAddress(Address address);
 	
-	public Address getAddressById(Integer addressId);
+	public Address getAddressById(Long addressId);
 }
